mlsec_improvise_adapt_overcome/src/requirements.txt
torch
torchvision
pillow
fastapi[all]
uvicorn[standard]
